API Move isAvailable() and getUnavailableMessage() to Method
Requires https://github.com/silverstripe/silverstripe-mfa/pull/82
