+++ src/TOTPAware.php
@@ -11,8 +11,8 @@ use SilverStripe\MFA\Store\StoreInterface;
 trait TOTPAware
 {
     /**
-     * Gets the encryption key to use from environment variables. This is generated by default on SilverStripe Platform
-     * and the Common Web Platform, but can be defined as a custom value if required.
+     * Gets the encryption key to use from environment variables. This is generated by default on the Common Web
+     * Platform, but can be defined as a custom value if required.
      *
      * @return string
      */
@@ -22,10 +22,11 @@ trait TOTPAware
     }
 
     /**
-     * Get an instance of the TOTP handler service
+     * Get an instance of the TOTP handler service. The secret must already be defined and set to the StoreInterface.
      *
      * @param StoreInterface $store
      * @return TOTPInterface
+     * @throws RuntimeException
      */
     protected function getTotp(StoreInterface $store): TOTPInterface
     {
